Refactor command registration to use shared LogisticsCommandTree
