+++ common/src/main/java/com/logistics/LogisticsPipe.java
@@ -6,6 +6,7 @@ import com.logistics.core.bootstrap.DomainBootstrap;
 import com.logistics.core.lib.resource.ResourceId;
 import com.logistics.pipe.modules.*;
 import com.logistics.pipe.network.NetDbg;
+import com.logistics.pipe.network.NetworkRegistry;
 import com.logistics.pipe.Pipe;
 import com.logistics.pipe.PipeApi;
 import com.logistics.pipe.PipeTypes;
@@ -23,6 +24,8 @@ import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -72,6 +75,16 @@ public final class LogisticsPipe extends LogisticsMod implements DomainBootstrap
         addCreativeTabEntries();
 
         LogisticsApi.Registry.transport(new PipeApi());
+    }
+
+    public static void onServerTick(MinecraftServer server) {
+        for (var level : server.getAllLevels()) {
+            NetworkRegistry.tickNetworks(level);
+        }
+    }
+
+    public static void onLevelUnload(ServerLevel level) {
+        NetworkRegistry.clearLevel(level);
     }
 
     private static void registerNetworkPackets() {
